handle the spaces and the lines

+++ controllers/productControllers.js
@@ -22,9 +22,9 @@ exports.postAddProduct = async(req, res) => {
     const image = req.file? req.file.filename : "";  
     const data = {title, category, desc, price, image}
     const slug = slugify(title, {lower: true, strict: true})
-    const categories = await categoryServices.getCategories();
-
+    let check = 1;
     try{
+        const categories = await categoryServices.getCategories();
         const existingProduct  = await productServices.getProduct(slug)
         if(existingProduct ){
             res.render('admin/add-product',{product: data, error: ['This product is already exists'], categories })
@@ -32,6 +32,7 @@ exports.postAddProduct = async(req, res) => {
         }
         const product = await productServices.addProduct(data);
         if(req.file && req.file.path){
+            check = 0;
             const productDir = path.join(__dirname, '..', 'public', 'uploads', 'products', product._id.toString())
             fs.mkdirSync(productDir, { recursive: true });
             const newPath = path.join(productDir, req.file.filename);
@@ -43,9 +44,9 @@ exports.postAddProduct = async(req, res) => {
             res.redirect('/admin/product');
         });
     }catch(error){
-        if (req.file) {
+        if (req.file && check == 1) {
             await fs.promises.unlink(path.join(__dirname, '../public', 'uploads', 'temp', req.file.filename));
-            }
+        }
         console.log(error)  
         res.render('admin/add-product',{product: data, error: ['there is something wrong in adding this product'], categories})
     }
@@ -111,7 +112,6 @@ const {title, category, desc, price, deleteImage} = req.body;
                 if(fs.existsSync(oldImagePath)){
                     fs.unlinkSync(oldImagePath);
                 }
-
             }
         }
         if(!req.file && deleteImage == true && existingProduct.image != "" ){
@@ -136,41 +136,40 @@ const {title, category, desc, price, deleteImage} = req.body;
 
 exports.deleteGallery = async(req, res) => {
     const {id, image} = req.params;
-        const galleryPath = path.join(__dirname,'../public/uploads', 'products', id, 'gallery', image)
-        if(fs.existsSync(galleryPath)){
-            fs.unlinkSync(galleryPath);
-            req.flash('success', 'Gallery image deleted successfully');
-            res.redirect(`/admin/product/edit-product/${id}`)
-        } else{
-            req.flash('error', 'this gallery image does not exist');
-            res.redirect(`/admin/product/edit-product/${id}`)
-        }
+    const galleryPath = path.join(__dirname,'../public/uploads', 'products', id, 'gallery', image)
+    if(fs.existsSync(galleryPath)){
+        fs.unlinkSync(galleryPath);
+        req.flash('success', 'Gallery image deleted successfully');
+        res.redirect(`/admin/product/edit-product/${id}`)
+    } else{
+        req.flash('error', 'this gallery image does not exist');
+        res.redirect(`/admin/product/edit-product/${id}`)
+    }
 }
 
 exports.deleteProduct = async(req, res) => {
-     const id = req.params.id;
-        try{
-            const existingProduct = await productServices.getProductID(id);
-            if(!existingProduct){
-                req.flash('error', 'this product is not exists')
-                return res.redirect('/admin/product/')
-            }
-            const deletedPath = path.join(__dirname, '../public/uploads/products');
-            if(!fs.existsSync(deletedPath)){
-                fs.mkdirSync(deletedPath, { recursive: true });
-            }
-            const productDir = path.join(deletedPath, id.toString());
-            if(fs.existsSync(productDir)){
-                fs.rmSync(productDir, { recursive: true, force: true });;
-            }
-            await productServices.deleteProduct(id);
-            req.flash('success', 'the product was deleted successfully')
-            res.redirect('/admin/product')
+    const id = req.params.id;
+    try{
+        const existingProduct = await productServices.getProductID(id);
+        if(!existingProduct){
+            req.flash('error', 'this product is not exists')
+            return res.redirect('/admin/product/')
         }
-        catch(error){
-            console.log(error)
-            req.flash('error', 'there is something wrong in deletting this product')
-            res.redirect('/admin/product/')
-    
+        const deletedPath = path.join(__dirname, '../public/uploads/products');
+        if(!fs.existsSync(deletedPath)){
+            fs.mkdirSync(deletedPath, { recursive: true });
         }
+        const productDir = path.join(deletedPath, id.toString());
+        if(fs.existsSync(productDir)){
+            fs.rmSync(productDir, { recursive: true, force: true });;
+        }
+        await productServices.deleteProduct(id);
+        req.flash('success', 'the product was deleted successfully')
+        res.redirect('/admin/product')
+    }
+    catch(error){
+        console.log(error)
+        req.flash('error', 'there is something wrong in deletting this product')
+        res.redirect('/admin/product/')
+    }
 }
